requirements.txt
fastapi==0.75.1
uvicorn==0.17.6
GitPython==3.1.27